requirements.txt
argh==0.26.2
arrow==0.15.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
beautifulsoup4==4.8.0
boto3==1.9.111
botocore==1.12.111
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.4
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-dotenv==1.4.2
django-dramatiq==0.8.0
django-extensions==2.0.7
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-phonenumber-field==2.2.0
django-redis-cache==2.1.0
django-storages==1.7.1
django-timezone-field==3.0
django-twilio==0.10.0
docutils==0.14
dramatiq==1.7.0
ez-setup==0.9
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
mock==3.0.5
model-mommy==1.6.0
more-itertools==7.2.0
mysqlclient==1.3.12
numpy==1.14.2
packaging==19.2
pandas==0.22.0
pathtools==0.1.2
phonenumbers==8.10.6
pika==1.1.0
Pillow==5.3.0
pluggy==0.13.0
prometheus-client==0.7.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
PySocks==1.6.8
pytest==5.2.0
python-crontab==2.3.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1.2
redis==3.3.8
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9.2
stripe==2.21.0
twilio==6.25.0
urllib3==1.24.1
watchdog==0.8.3
watchdog-gevent==0.1.0
wcwidth==0.1.7
whitenoise==4.1.4
xlrd==1.1.0
zipp==0.6.0
